requirements.txt
Django==3.1.6
mysqlclient==2.0.3
django-mysql==3.11.1
django-admin-rangefilter==0.6.4